create BinaryExactVersionHintUrlInfo trait
