+++ src/driver_management.rs
@@ -11,6 +11,7 @@ pub mod chromedriver_info;
 pub mod installation_info;
 pub mod url_info;
 pub mod verification_info;
+mod binary_exact_version_hint_url_info;
 
 pub trait WebdriverInfo:
     WebdriverUrlInfo + WebdriverInstallationInfo + WebdriverVerificationInfo + Sync
